Added slow queries log and slack notifications
It's now possible to use two new configuration options, either via command line
switches or configuration file:

- `slow_threshold` (default 60)

If a query has been running for longer than this threshold, it is highlighted
in the list of running queries and it is also logged to a `pgtop.log` file.

- `slack_webhook` (no default)

When a slack webhook URL is provided, if there's any queries whose runtime
exceeds the slow threshold, pgtop will send a slack notification providing all
the information about the slow query, host, database, user, query ID and the
corresponding SQL of the query.